Move home link to appname
Previously, there was a link to the home on the top-right of each page.
I've removed this link and instead, you can go home by clicking on the
app name on the top left of every page.  This is standand, saves space
on the navbar and simplifies the app visually.